% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlbatch.R
\name{sql_batch}
\alias{sql_batch}
\title{Run all SQL statement}
\usage{
sql_batch(con, sql = NA, file = NA, clean = FALSE)
}
\arguments{
\item{con}{Connection to a Database}

\item{sql}{Vector of strings containing SQL statements 
(either single statements or multiple statements separated by a semicolon)}

\item{file}{Vector of filename containing SQL statement}

\item{clean}{Cleaning SQL statement before execution?
(dropping of empty line, dropping of comments, ...)}
}
\value{
data (if SQL contains a SELECT statement) otherwise NA
}
\description{
Runs all SQL statements in a vector of files or in a vectot of strings.
If sql contains multiple statements separated by a semicolon
all SQL statements are run in serial.
}
\examples{
\dontrun{
sql_run("SELECT * \nFROM mytable")
}
}
