% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlbatch.R
\name{sql_read_file}
\alias{sql_read_file}
\title{Read SQL from a file}
\usage{
sql_read_file(file)
}
\arguments{
\item{file}{path and filename}
}
\value{
SQL statement
}
\description{
Reading SQL statements from a text-file. 
The file may contain a single SQL or multiple SQL separated by semicolon
}
\examples{
\dontrun{
sql <- sql_read_file("C:/temp/demo.sql")
}
}
