% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlbatch.R
\name{sql_run}
\alias{sql_run}
\title{Run SQL statement}
\usage{
sql_run(con, sql = NA, file = NA, clean = FALSE)
}
\arguments{
\item{con}{Connection to a Database}

\item{sql}{SQL statement}

\item{file}{Filename containing SQL statement}

\item{clean}{Cleaning SQL statement before execution?
(dropping of empty line, dropping of comments, ...)}
}
\value{
data (if SQL contains a SELECT statement)
}
\description{
Runs all SQL statements in a file or in sql.
If sql contains multiple statements separated by a semicolon
all SQL statements are run in serial.
}
\examples{
\dontrun{
sql_run("SELECT * \nFROM mytable")
}
}
