% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlbatch.R
\name{sql_show}
\alias{sql_show}
\title{Print SQL statement to console.}
\usage{
sql_show(sql)
}
\arguments{
\item{sql}{SQL statement}
}
\value{
SQL statement
}
\description{
Print SQL statement to console.
}
\examples{
sql_show("SELECT * \nFROM mytable")
}
