% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlbatch.R
\name{sql_split}
\alias{sql_split}
\title{Split SQL statement}
\usage{
sql_split(sql)
}
\arguments{
\item{sql}{SQL statement}
}
\value{
Vector of SQL statements
}
\description{
Splits SQL containing multiple statements separated by a semicolon
into a vector of single SQL statements.
}
\examples{
sql_show("SELECT * \nFROM mytable")
}
