% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlbatch.R
\name{sql_clean}
\alias{sql_clean}
\title{Clean SQL statement}
\usage{
sql_clean(sql)
}
\arguments{
\item{sql}{SQL statement}
}
\value{
SQL statement
}
\description{
Clean SQL statement
}
\examples{
sql_clean("\n\n-- Test\nSELECT * FROM mytable")
}
